+++ src/main/java/com/team9/anicare/information/service/InformationService.java
@@ -10,8 +10,8 @@ import com.team9.anicare.information.dto.InformationRequestDTO;
 import com.team9.anicare.information.dto.InformationResponseDTO;
 import com.team9.anicare.information.model.Information;
 import com.team9.anicare.information.repository.InformationRepository;
-import com.team9.anicare.species.model.Breed;
-import com.team9.anicare.species.repository.BreedRepository;
+import com.team9.anicare.animal.model.Breed;
+import com.team9.anicare.animal.repository.BreedRepository;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
